moved the text util 'replaceClassName' into decorate.js and imported into block
