+++ libs/blocks/text/text.js
@@ -1,4 +1,4 @@
-import { decorateBlockBg, decorateBlockText, getBlockSize } from '../../utils/decorate.js';
+import { decorateBlockBg, decorateBlockText, getBlockSize, replaceClassName } from '../../utils/decorate.js';
 
 // size: [heading, body, ...detail]
 const blockTypeSizes = {
@@ -21,41 +21,6 @@ const blockTypeSizes = {
     xlarge: ['xxl', 'l', 'xl'],
   },
 };
-
-function findElementWithClassStartingOrEndingWith(element, searchString) {
-  var classes = element.classList;
-  var foundClass = "";
-  for (var i = 0; i < classes.length; i++) {
-    if (classes[i].startsWith(searchString) || classes[i].endsWith(searchString)) {
-      foundClass = classes[i];
-      break;
-    }
-  }
-  if (foundClass) {
-    var parts = foundClass.split("-");
-    var newStr = parts[1] + "-" + parts[0];
-    return { element: element, class: foundClass, newStr: newStr };
-  } else {
-    return null;
-  }
-}
-
-function replaceClassName(el, str) {
-  const foundEl = findElementWithClassStartingOrEndingWith(el, str);
-  if (foundEl) {
-    const findClass = str.slice(1) + str[0];
-    const els = foundEl.element.querySelectorAll(`[class^="${findClass}"]`);
-    if (!els) return;
-    [...els].forEach( (e, i) => {
-      for (let i = 0; i < e.classList.length; i++) {
-        const className = e.classList[i];
-        if (className.startsWith(findClass)) {
-          e.classList.replace(className, foundEl.newStr);
-        }
-      }
-    });
-  };
-}
 
 async function applyOverrides(el) {
   const overrides = ['-heading', '-body', '-detail'];
